vhost: start vring only on first kickfd event
Spec said that vrings are started when they receive a kick event. It
doesn't say however that this kickfd event will also have no other
purpose but to start the vring (as opposed to also signalling valid
buffers in the vring).

Do not attempt to handle vring events and only start vring if we receive
a kickfd event and vring is stopped.

Signed-off-by: Evgeny Yakovlev <eyakovlev3@gmail.com>